+++ script/core/completion/completion.lua
@@ -1354,20 +1354,6 @@ local function getCallEnumsAndFuncs(source, index, oop, call)
         end
     end
     if source.type == 'doc.type.function' then
-        --[[
-        always use literal index, that is:
-        ```
-        ---@class Class
-        ---@field f(x: number, y: boolean)
-        local c
-
-        c.f(1, true) -- correct
-        c:f(1, true) -- also correct
-        ```
-        --]]
-        if oop then
-            index = index - 1
-        end
         local arg = source.args[index]
         if arg and arg.extends then
             return pushCallEnumsAndFuncs(vm.getDefs(arg.extends))
@@ -1959,6 +1945,7 @@ local function makeCache(uri, position, results)
     cache.position= position
     cache.word    = word:lower()
     cache.length  = #word
+    cache.uri     = uri
 end
 
 local function isValidCache(word, result)
@@ -1981,6 +1968,9 @@ end
 local function getCache(uri, position)
     local cache = workspace.getCache 'completion'
     if not cache.results then
+        return nil
+    end
+    if cache.uri ~= uri then
         return nil
     end
     local text  = files.getText(uri)