+++ script/core/definition.lua
@@ -4,6 +4,7 @@ local files      = require 'files'
 local vm         = require 'vm'
 local findSource = require 'core.find-source'
 local guide      = require 'parser.guide'
+local rpath      = require 'workspace.require-path'
 
 local function sortResults(results)
     -- 先按照顺序排序
@@ -74,7 +75,7 @@ local function checkRequire(source, offset)
         return nil
     end
     if     libName == 'require' then
-        return workspace.findUrisByRequirePath(literal)
+        return rpath.findUrisByRequirePath(literal)
     elseif libName == 'dofile'
     or     libName == 'loadfile' then
         return workspace.findUrisByFilePath(literal)
@@ -155,6 +156,9 @@ return function (uri, offset)
         or  src.type == 'setindex'
         or  src.type == 'tableindex' then
             src = src.index
+            if not src then
+                goto CONTINUE
+            end
             if not guide.isLiteral(src) then
                 goto CONTINUE
             end