+++ changelog.md
@@ -1,6 +1,39 @@
 # changelog
 
+## 2.5.5
+`2021-12-16`
+* `FIX` dose not work in VSCode
+
+## 2.5.4
+`2021-12-16`
+* `FIX` [#847](https://github.com/sumneko/lua-language-server/issues/847)
+* `FIX` [#848](https://github.com/sumneko/lua-language-server/issues/848)
+* `FIX` completion: incorrect cache
+* `FIX` hover: always view string
+
+## 2.5.3
+`2021-12-6`
+* `FIX` [#842](https://github.com/sumneko/lua-language-server/issues/844)
+* `FIX` [#844](https://github.com/sumneko/lua-language-server/issues/844)
+
+## 2.5.2
+`2021-12-2`
+* `FIX` [#815](https://github.com/sumneko/lua-language-server/issues/815)
+* `FIX` [#825](https://github.com/sumneko/lua-language-server/issues/825)
+* `FIX` [#826](https://github.com/sumneko/lua-language-server/issues/826)
+* `FIX` [#827](https://github.com/sumneko/lua-language-server/issues/827)
+* `FIX` [#831](https://github.com/sumneko/lua-language-server/issues/831)
+* `FIX` [#837](https://github.com/sumneko/lua-language-server/issues/837)
+* `FIX` [#838](https://github.com/sumneko/lua-language-server/issues/838)
+* `FIX` postfix
+* `FIX` runtime errors
+
+## 2.5.1
+`2021-11-29`
+* `FIX` incorrect syntax error
+
 ## 2.5.0
+`2021-11-29`
 * `NEW` settings:
   + `Lua.runtime.pathStrict`: not check subdirectories when using `runtime.path`
   + `Lua.hint.await`: display `await` when calling a function marked as async